backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
pdfplumber==0.10.3
Pillow==10.1.0
deepl==1.15.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
beautifulsoup4==4.12.2
transformers>=4.30.0
torch>=2.0.0

